Metric Shit ton of things.
Reworked almost all of game to make the win method actually work. I replaced most of the person variables with playerarray.length. I also refactored game to make use of the class wide Player array list so removing players would be possible. Renamed the player array to playerarray for ease of use. Worked in almost everyclass to do this. Sorry not sorry. Also made it so that when a player is removed all of their property is given back to the banker and it gets its isBought variable set to false so it can be rebought. Will still need to make it so that all the property can go to another player. As of right now it can only go to the bank.